asset-info renamed to sketch-info
