--- src/components/sketch/asset-info/index.js
@@ -1,3 +1,0 @@
-import AssetInfo from "./assetinfocontroller"
-
-export default AssetInfo
