+++ src/components/sketch/sketch-info/index.js
@@ -1,0 +1,3 @@
+import SketchInfo from "./sketchinfocontroller"
+
+export default SketchInfo
