+++ src/components/sketch/sketch-info/sketchinfo.scss
@@ -1,3 +1,3 @@
-.asset-info {
+.sketch-info {
     padding: 2em;
 }
